+++ src/components/public/gnb/Profile/Atoms/Authenticated.tsx
@@ -1,15 +1,17 @@
 "use client"
 
 import { signOut } from "next-auth/react"
+import Image from "next/image"
 import { useRouter } from "next/navigation"
 
 import { useRef, useState } from "react"
 
-import Profile from "@/components/public/img/Profile"
 import ROUTE from "@/constants/route"
+import useGetUserData from "@/hooks/useGetUserData"
 import useOutsideClick from "@/util/useOutsideClick"
+import ProfileLargeEditIMG from "@public/img/profile_large_edit.png"
 
-const Authenticated = ({ profileImg }: { profileImg: string | undefined | null }) => {
+const Authenticated = ({ token }: { token: string }) => {
   const router = useRouter()
   const [isOpen, setIsOpen] = useState(false)
   const dropdownRef = useRef<HTMLDivElement>(null)
@@ -22,14 +24,22 @@ const Authenticated = ({ profileImg }: { profileImg: string | undefined | null }
     return setIsOpen(!isOpen)
   }
 
+  const { data } = useGetUserData(token)
+
   return (
     <div ref={dropdownRef} className="mt-1 md:mt-2">
       <button type="button" onClick={handleToggle} aria-label="프로필 메뉴 열기">
-        <Profile state="largeDefault" className="h-[40px] w-[40px]" profileImg={profileImg} />
+        <Image
+          src={data?.image || ProfileLargeEditIMG}
+          width={40}
+          height={40}
+          className="h-10 w-10 rounded-full"
+          alt="유저 이미지"
+        />
       </button>
 
       {isOpen && (
-        <div className="absolute right-0 top-[62px] flex h-[70px] w-[150px] flex-col rounded-lg bg-white shadow-xl md:top-[66px]">
+        <div className="absolute right-5 top-[62px] flex h-[70px] w-[150px] flex-col rounded-lg bg-white shadow-xl md:top-[66px]">
           <button
             type="button"
             className="flex h-1/2 w-full transform items-center justify-center rounded-lg text-center text-orange-600 transition-all delay-[10ms] duration-150 ease-in-out"
