+++ src/components/public/gnb/Profile/Profile.tsx
@@ -5,11 +5,7 @@ import { useSession } from "next-auth/react"
 import Authenticated from "@/components/public/gnb/Profile/Atoms/Authenticated"
 import NotAuthenticated from "@/components/public/gnb/Profile/Atoms/NotAuthenticated"
 
-interface IProfileComponentProps {
-  profileImg: string | undefined | null
-}
-
-const Profile = ({ profileImg }: IProfileComponentProps) => {
+const Profile = () => {
   const session = useSession()
   const token = session.data?.accessToken
   const isNotToken = !token && session.status !== "loading"
@@ -17,7 +13,7 @@ const Profile = ({ profileImg }: IProfileComponentProps) => {
   return (
     <>
       {isNotToken && <NotAuthenticated />}
-      {token && <Authenticated profileImg={profileImg} />}
+      {token && <Authenticated token={session.data?.accessToken || ""} />}
     </>
   )
 }
